+++ bitcoin/src/blockdata/transaction.rs
@@ -118,9 +118,8 @@ impl OutPoint {
     /// # Examples
     ///
     /// ```rust
-    /// use bitcoin::consensus::params;
     /// use bitcoin::constants::genesis_block;
-    /// use bitcoin::Network;
+    /// use bitcoin::{params, Network};
     ///
     /// let block = genesis_block(&params::MAINNET);
     /// let tx = &block.txdata[0];
